fix: get data pemeriksaan ke database

+++ app/src/main/java/com/mybaseprojectandroid/model/Pemeriksaan.kt
@@ -5,5 +5,5 @@ data class Pemeriksaan(
     val idUser: String? = null,
     val tanggal: String? = null,
     val jenis: String? = null,
-    val nilai: Int? = null
+    val nilai: Float? = null
 )
